# Conflicts:
#	e2e/cli/autoscaling/test_instance_kill_by_timout.py
#	e2e/cli/autoscaling/test_kill_node_manually.py
#	e2e/cli/autoscaling/test_node_reassign.py
#	e2e/cli/autoscaling/test_on_demand_pipe_run.py
#	e2e/cli/autoscaling/test_pipline_run_sync.py
#	e2e/cli/autoscaling/test_spot_pipe_run.py
#	e2e/cli/autoscaling/test_start_stop_pipe.py
#	e2e/cli/autoscaling/test_stop_pipe_before_labeling.py
#	e2e/cli/autoscaling/test_terminate_instance_before_registration.py
#	e2e/cli/autoscaling/test_terminate_instance_during_pipe_work.py
#	e2e/cli/autoscaling/test_terminate_node_during_pipe_work.py
#	e2e/cli/autoscaling/test_two_pipes_simultaneously.py
#	e2e/cli/autoscaling/test_unused_instance_termination.py

+++ deploy/contents/install/app/install.sh
@@ -258,6 +258,10 @@ CP_TP_KUBE_NODE_NAME=${CP_TP_KUBE_NODE_NAME:-$KUBE_MASTER_NODE_NAME}
 print_info "-> Assigning cloud-pipeline/cp-tinyproxy to $CP_TP_KUBE_NODE_NAME"
 kubectl label nodes "$CP_TP_KUBE_NODE_NAME" cloud-pipeline/cp-tinyproxy="true" --overwrite
 
+# Allow to schedule policy-manager service to the master
+CP_POLICY_MANAGER_KUBE_NODE_NAME=${CP_POLICY_MANAGER_KUBE_NODE_NAME:-$KUBE_MASTER_NODE_NAME}
+print_info "-> Assigning cloud-pipeline/cp-run-policy-manager to $CP_POLICY_MANAGER_KUBE_NODE_NAME"
+kubectl label nodes "$CP_POLICY_MANAGER_KUBE_NODE_NAME" cloud-pipeline/cp-run-policy-manager="true" --overwrite
 
 echo
 
@@ -1180,15 +1184,30 @@ if is_service_requested cp-billing-srv; then
     echo
 fi
 
-
 # OOM reporter - monitor and report OOM related events for each pipeline run
-print_ok "[Starting OOM reporter daemonset deployment]"
+if is_service_requested cp-oom-reporter; then
+  print_ok "[Starting OOM reporter daemonset deployment]"
 
-print_info "-> Deleting existing instance of OOM reporter daemonset"
-kubectl delete daemonset cp-oom-reporter
+  print_info "-> Deleting existing instance of OOM reporter daemonset"
+  kubectl delete daemonset cp-oom-reporter
+  if is_install_requested; then
+    print_info "-> Deploying OOM reporter daemonset"
+    create_kube_resource $K8S_SPECS_HOME/cp-oom-reporter/cp-oom-reporter.yaml
+  fi
+fi
 
-print_info "-> Deploying OOM reporter daemonset"
-create_kube_resource $K8S_SPECS_HOME/cp-oom-reporter/cp-oom-reporter.yaml
+# Run-policy manager - monitor and manage network policies to implement restrictions on inter-run connections
+if is_service_requested cp-run-policy-manager; then
+  print_ok "[Starting run-policy manager deployment]"
+  print_info "-> Deleting existing instance of run-policy manager"
+  delete_deployment_and_service "cp-run-policy-manager" \
+                                  "/opt/run-policy-manager"
+  if is_install_requested; then
+    print_info "-> Deploying run-policy manager"
+    create_kube_resource $K8S_SPECS_HOME/cp-run-policy-manager/cp-run-policy-manager-dpl.yaml
+    wait_for_deployment "cp-run-policy-manager"
+  fi
+fi
 
 print_ok "Installation done"
 echo -e $CP_INSTALL_SUMMARY
